pc配置中心-产品商城100%

+++ core/src/main/java/com/smart/home/modules/product/service/ProductShopService.java
@@ -1,9 +1,13 @@
 package com.smart.home.modules.product.service;
 
 import com.github.pagehelper.PageHelper;
+import com.smart.home.common.exception.DuplicateDataException;
+import com.smart.home.common.exception.ServiceException;
+import com.smart.home.modules.product.dao.ProductMapper;
 import com.smart.home.modules.product.dao.ProductShopMapper;
 import com.smart.home.modules.product.entity.ProductShop;
 import com.smart.home.modules.product.entity.ProductShopExample;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -19,13 +23,28 @@ public class ProductShopService {
 
     @Resource
     ProductShopMapper productShopMapper;
+    @Resource
+    ProductMapper productMapper;
 
     public int create(ProductShop productShop) {
+        // 检查唯一性
+        ProductShopExample example = new ProductShopExample();
+        example.createCriteria().andShopNameEqualTo(productShop.getShopName());
+        if (productShopMapper.countByExample(example) > 0) {
+            throw new DuplicateDataException("该商城已经存在");
+        }
         productShop.setCreatedTime(new Date());
+        productShop.setRevision(0);
         return productShopMapper.insertSelective(productShop);
     }
 
     public int update(ProductShop productShop) {
+        // 检查唯一性
+        ProductShopExample example = new ProductShopExample();
+        example.createCriteria().andShopNameEqualTo(productShop.getShopName()).andIdNotEqualTo(productShop.getId());
+        if (productShopMapper.countByExample(example) > 0) {
+            throw new DuplicateDataException("该商城已经存在");
+        }
         productShop.setUpdatedTime(new Date());
         return productShopMapper.updateByPrimaryKeySelective(productShop);
     }
@@ -37,6 +56,10 @@ public class ProductShopService {
     @Transactional(rollbackFor = RuntimeException.class)
     public void delete(List<Long> idList) {
         for (Long id : idList) {
+            // 检查有没有挂钩产品了
+            if (productMapper.countByShopId(id.intValue()) > 0) {
+                throw new ServiceException("已经有关联的产品，不能进行删除");
+            }
             productShopMapper.deleteByPrimaryKey(id.intValue());
         }
     }
@@ -45,7 +68,9 @@ public class ProductShopService {
         PageHelper.startPage(pageNum, pageSize);
         ProductShopExample example = new ProductShopExample();
         ProductShopExample.Criteria criteria = example.createCriteria();
-        // TODO 按需根据字段查询
+        if (StringUtils.isNotBlank(productShop.getShopName())) {
+            criteria.andShopNameLike(productShop.getShopName());
+        }
         return productShopMapper.selectByExample(example);
     }
 
@@ -54,4 +79,8 @@ public class ProductShopService {
         return productShop;
     }
 
+    public List<ProductShop> queryAllValid() {
+        ProductShopExample example = new ProductShopExample();
+        return productShopMapper.selectByExample(example);
+    }
 }
